requirements.txt
chromadb==1.0.6
langchain==0.3.24
PyGithub==2.6.1
PyGithub==2.6.1
python_docx==1.1.2
Requests==2.32.3
sentence_transformers==4.1.0
streamlit==1.44.1
